requirements.txt
python-dotenv
google-generativeai
PyPDF2
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu